+++ packages/pose-core/CHANGELOG.md
@@ -2,11 +2,17 @@
 
 Pose Core adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.0.2] 2018-08-20
+
+### Fixed
+
+- Fixing bug in pose unsetting priority.
+
 ## [1.0.1] 2018-08-19
 
 ### Fixed
 
-Updating `unset` logic.
+- Updating `unset` logic.
 
 ## [1.0.0] 2018-08-19